+++ e2e/tests-dfx/frontend.bash
@@ -81,3 +81,20 @@ teardown() {
     assert_command curl -vv http://localhost:"$PORT"/index.js?canisterId="$ID"
     assert_match "< x-key: x-value"
 }
+
+@test "dfx uses a custom build command if one is provided" {
+    jq '.canisters.e2e_project_frontend.source = ["dist/e2e_project_frontend/"]' dfx.json | sponge dfx.json
+    jq '.canisters.e2e_project_frontend.build = ["npm run custom-build"]' dfx.json | sponge dfx.json
+    jq '.scripts["custom-build"] = "mkdir -p ./dist/e2e_project_frontend/assets/ && cp -r ./src/e2e_project_frontend/assets/* ./dist/e2e_project_frontend"' package.json | sponge package.json
+
+    dfx_start
+    dfx canister create --all
+    dfx build
+    dfx canister install --all
+
+    ID=$(dfx canister id e2e_project_frontend)
+    PORT=$(get_webserver_port)
+
+    assert_command curl -vv http://localhost:"$PORT"/sample-asset.txt?canisterId="$ID"
+    assert_match "This is a sample asset!"
+}