+++ CHANGELOG.md
@@ -4,6 +4,21 @@
 
 ## DFX
 
+### feat: add optional custom build command for asset canisters
+
+The custom build command can be set in `dfx.json` the same way it is set for `custom` type canisters. If the command is not provided, DFX will fallback to the default `npm run build` command.
+
+```json
+{
+  "canisters": {
+    "ui": {
+      "type": "assets",
+      "build": ["<custom build command>"]
+    }
+  }
+}
+```
+
 ### fix: motoko canisters can import other canisters with service constructor
 
 After specific canister builder output wasm and candid file, `dfx` will do some post processing on the candid file.
@@ -146,8 +161,10 @@ Updated ic-ref to 0.0.1-a9f73dba
 
 ### Replica
 
-Updated replica to elected commit b3b00ba59c366384e3e0cd53a69457e9053ec987.
+Updated replica to elected commit 794fc5b9341fa8f6a0e8f219201c35f0b5727ab9.
 This incorporates the following executed proposals:
+- [122617](https://dashboard.internetcomputer.org/proposal/122617)
+- [122615](https://dashboard.internetcomputer.org/proposal/122615)
 - [122529](https://dashboard.internetcomputer.org/proposal/122529)
 - [122284](https://dashboard.internetcomputer.org/proposal/122284)
 - [122198](https://dashboard.internetcomputer.org/proposal/122198)